chore: identify all unreleased related images for an FBC fragment
The set of related images will only continue to grow for a FBC fragment
as bundles are added to the packages' channels. Once fragments are
released and become part of the target index, we no longer need to
consider the validity of these related images. Therefore, we want to be
able to determine the set of unique unreleased related images for an FBC
fragment so that we know the minimal set of pullspecs that might need to
be further inspected and/or verified.

Signed-off-by: arewm <arewm@users.noreply.github.com>