fix: improve memory utilization when extracting differential metadata
Memory consumption when slurping in the fbc and index metadata was
necessitating bumping the memory for pods. We can instead do this set
subtraction in a more efficient way directly in bash.

By changing extract_unique_bundles_from_catalog, this commit also
simplifies the postprocessing of results as returned during the
extracting of differential fbc metadata.

Signed-off-by: arewm <arewm@users.noreply.github.com>