chore: remove coverity checks rom build pipelines
Coverity is SAST scanner that Konflux only needs in order to scan C/C++
code. Since there is no C/C++ code in the konflux-test image we really
don't need it.  Additionally, its the longest-running check in the
konflux-test build pipeline, usually taking about 20 minutes.  Because
of this it makes sense to remove Coverity entirely in order to speed up
our builds

Signed-off-by: Ryan Cole <rcyoalne@gmail.com>